Iterator behavior changed to support compatibility with <algorithms>
I tried to mimic the STL's iterators as their end() points to one past
of the last element. This way we can support compatibility with
algorithms from the <algorithms> header. Also, writing loops are so much
easier as there is no need for an extra control statement for detecting
the last node. In addition, the range-for works properly :)